device/moto/stingray: whisper: speaker dock detection
A speaker dock is a basic low-end dock with the speaker built into the dock. To
the apps it appears as though a headset is inserted into the basic dock.

We don't need the last byte to detect a LE dock. The byte could be reused to
pass other properties.

Change-Id: If88794535e836442471c509b90a045c2175bad2f
Signed-off-by: Praveen Bharathi <pbharathi@motorola.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>