Audio glitches removal.
When starting and stopping audio capture while playback is active,
gaps can be heard in audio playback.This is because the input stream
configuration can take several ms during which the output stream write()
function can be waiting for the global mutex hold by doRouting_l().

The fix consists in not acquiring the global mutex in write() or read() unless
a change is required in output or input configuration.

Additional improvements:
- Make sure that the EC/NS state and output or input sampling rate changes
are an atomic operation that can not take place concurrently with a read or write.
- Do not change input path if the input stream is not active
- Close input driver when entering standby to allow other input streams activation
- Do not set kernel input or output device if same as previous
- Do not update kernel gains if same as previous.

Change-Id: I959086adf29037b539893ea520123cef1dc441c5