Make tty2ttyd setup Netlink socket credentials
The kernel was recently updated to filter Netlink events from going to
entities with improper or no socket credentials.  The RIL was already
updated to deal with this, so that it would detect when the MDM6600
dies.  tty2ttyd also needs to know when the MDM6600 or Wrigley dies,
but it's not receiving the events due to the lack of socket
credentials.  Add the setting of them just like the RIL does.

Change-Id: If162ff40e7521865c2c6980233f8829226861767