CDMA RILv3.12
-- Use "SIM" radio states instead of "NV" for Everest.
This enables use of SIM_NOT_READY, which will prevent FW from
accessing SIM when the SIM is initializing after power up or recovery.

Also includes a fix to not require a CSIM/RUIM app in CDMA mode when
setting the ICC "READY" radio state. This is necessary because of
Verizon CDMA testing that uses a USIM-only SIM card. Don't think
that removing this check will negatively affect any other product.

-- Handle service configuration change/unsol from BP (if BP
supports). This occurs when user inserts or removes a SIM.

Change-Id: I7c511d97a90607b0cbef75740b2cbd44cb2f853d