lights: Enable notification LED logging
It is currently impossible to debug notification LED issues
because there is no logging enabled.  We have no way of knowing if
the driver failed to change the LED state, or if Android never even
requested the new LED state.  This print will allow us to see
whether Android has in fact requested a new notification LED state.

Needed for Bug 4643166.

Change-Id: Id2efc41fa6e1d5e5b9843d1945ad146f72184464
Signed-off-by: Greg Meiste <w30289@motorola.com>