Add logcatd/logcatd-blan for MDL logging
logcatd serves the logcat and kernel logs to the MDL logger.  It
does so strictly over TCP/IP, to avoid any dependency on ADB.  ADB is
not reliable enough for automation, and adds an unnecessary dependency
on the logger tool.

logcatd-blan is a wrapper around logcatd to watch for the Moto BLAN
iface and fork-exec a logcatd child when it's up.

Bug: 4993645
Change-Id: I9aa8035c98cfdf8e02a785f2f1f6e082594d0246